+++ docease/src/main/java/com/example/docease/repositories/AuditLogRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface AuditLogRepository extends JpaRepository<AuditLog, Long> {
+public interface AuditLogRepository extends JpaRepository<AuditLog, Integer> {
     List<AuditLog> findByUser(User user);
 
 }
